refactoring and review comments
